Application Status Colors and Grid Sorting functionality
Added coloring to the main grid based on the different application statuses. If you have applied status and applied more than 60 days ago it will show red. If it was 30-59 days ago, it will be yellow. If it was less than 30 days ago, it will be lime. Denied status will show red. R1 interview will show sky blue. Technical interview will be purple.

Also added sorting for the grid upon column header click. There are no elements to show this as possible, but seems like it works properly so far.